// src/schemas/public/UserAddresses.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { UsersId } from './Users';

/** Identifier type for public.user_addresses */
export type UserAddressesId = number & { __brand: 'UserAddressesId' };

/** Represents the table public.user_addresses */
export default interface UserAddresses {
  id: UserAddressesId;

  user_id: UsersId;

  name: string;

  address: string;

  city: string;

  state: string;

  postal_code: string;

  country: string;
}

/** Represents the initializer for the table public.user_addresses */
export interface UserAddressesInitializer {
  /** Default value: nextval('user_addresses_id_seq'::regclass) */
  id?: UserAddressesId;

  user_id: UsersId;

  name: string;

  address: string;

  city: string;

  state: string;

  postal_code: string;

  country: string;
}

/** Represents the mutator for the table public.user_addresses */
export interface UserAddressesMutator {
  id?: UserAddressesId;

  user_id?: UsersId;

  name?: string;

  address?: string;

  city?: string;

  state?: string;

  postal_code?: string;

  country?: string;
}